Add more interactivity to the JS project
Instead of using the withConfig gl-wrapper module function for setting the
viewport, this is done in the draw method call instead. Change how the
lighting for the potential surface is calculated. Fix Quad recycling.
Normalize the wave function using a checkbox instead of automatically
doing it when the time step is an imaginary value.
When placing a new wave packet periodic boundary conditions are obeyed.
Add time-dependent preset potential options. Create controls for changing
the grid size. Include sqrt as a function that can be parsed by the text
entries. Sum reduction of a texture is done through GLSL shaders instead
of summing its typed array. Add initializeDefaultKineticEnergy for computing
the exponential of the kinetic energy term in JS instead of GLSL.
Implement touch interaction. Begin creating a 3D volume render.