Implement many new features
The primary view canvas (not the simulation domain) is no longer necessarily
a square; its aspect ratio is determined by the client screen dimensions.
Loading an image for the potential should now be done when the image is ready
instead of waiting for a second to check if the image has been loaded.
Images are cropped instead of allotting empty space whenever the aspect ratio
of the image does not match the simulation's.
Sketching the potential with the mouse should now be
more continuous and flowing.
Add stats to the left div for displaying the position of the mouse cursor and
the initial wave packet momentum.
Modify the upside down Gaussian potential preset with controls for changing
the radius of rotation.
When changing the grid size, rather than recycling the original quads and
substituting it with new ones of the desired size, a method is instead
called that only modifies the quad's frame buffer and resizes its texture.
For sum reduction of a source texture, the textures used when conducting the
summation uses the source format except for the last single pixel texture,
which uses RGBA32F.
Found a fix for an issue that caused the simulation to error out with
NaN values whenever normalize wave function is set to true and then some other
settings are edited such as the grid size: apparently removing the calls to
gl.ActiveTexture or gl.bindTexture inside the method asFloat32Array fixed
this issue.
Move shaders into subdirectories instead of placing them in one large
directory.
Also begin creating shaders for implementing the split step method for the
Dirac equation.