Create window for viewing wave function in momentum space
Split the canvas initialization from gl-wrappers in to its own
script. The first quad is now created in gl-wrappers, where it
must be imported from. Add a slider for the harmonic oscillator
preset potential. Begin implementing new function shapes for
wavepacket initialization. Continue implementing the volume render.
Correct a few typos in some comments and change the wording of
some of the UI.