Modify how summations are done and fix issue with planar slice render
Only two summation quads are used, which are reset to use a different
texture size at each iteration of the loop. Originally
log2(side length of source texture) quads were used, which caused
the program to run out of available texture units to use, making the
final summation result incorrect. The performance loss when reseting
the summation quad's textures at each iteration appears to be
negligible.

For the offset of the xz and yz planes, taking the additive inverse
of these values appears to correctly align these slices with the
volume data.

Also in the gradient function for the volume-render module, integer
parameters are passed as arguments instead of IScalar. The IScalar
objects are instead created in the body of the functino when calling
the destination quad's draw method.