New controls for modifying volume render visualization
When colour phase checkbox is unselected, the wave function should
now appear in a different colour than the potential. Modify
wave function brightness calculation when 1 / |psi| is selected
so that tube vortices and other structures are visible when
show phase is unchecked.

Fix an issue when switching grid dimensions that caused the program
to run out of texture memory, by resetting the volume render data
quads instead of replacing them whenever the dimensions are changed.

Modify fft3d shader so that the 3D simulation works with WebGL1.

Begin a new fft iteration implementation that operates specifically
on square texture sizes.