New volume render controls
Can change the total number of slices and the slice side width.
Also modify the texture formats of some of the quads used
internally by the volume render. Change default positions of
the colour and alpha slider controls. In the display fragment
shader prototype various filtering schemes that are meant to
smooth out numerical artifacts that appear in the final render.

For both the 2D and 3D simulation add hyperlinks to the source
code. For the 3D simulation add text display at the bottom left
portion of the canvas showing the position of the cursor in
simulation space.

Get the 2D FFT square iteration shader to work, which operates
specifically on textures with equal side lengths. Implement
a 3D FFT cube version as well, which is meant for textures
storing 3D arrays of data whose side lengths are equal.
For the 3D FFT set useCosTable to false, which slightly
increases performance.