Fix numerous issues and add new features for the Dirac 2D simulation
Add mouse selection for erasing the potential.
The time step modifies c*dt/dx instead of the time step directly,
where c is the speed of light, dt is the time step, and dx is the
spatial step size. Implement the pseudoscalar visualization and fix how
the pseudocurrent is computed. Add checkbox selection where the
wave function can either be initialized as a product of a real-valued
Gaussian with a plane wave, or as a superposition of plane waves.
When changing the proportion of positive E in the initial wave packet,
the corresponding amount of negative E is displayed as well.
Sliders for controlling the spin orientation of the initial wave function
are grouped into two sets of three, where the first group modifies the
spin orientation of the positive E solutions while the second controls
the orientation for the negative E, instead of originally having six
completely separate sliders. Add visualizations for the E and B fields.
Create a dropdown for selecting preset potentials.

Begin implementation of controls where the wave function can be
initialized with the mouse when the 3D view is selected.

Refactor location of classes and other code structure.
Fix an issue with how the spin was computed for those orientations that
are confined to the z direction.