Fix problems that occur on mobile devices
Change how the HTML pages are formatted when the screen
orientation is on portrait instead of landscape.
Add default min and mag filter: for mobile devices,
this defaults to gl.NEAREST because Safari on IOS does not
suport floating point linear textures; on other devices
use gl.LINEAR as before.
For the 3D simulation add basic touch controls for rotating
the simulation. Implementing texture sum reduction for
gl.NEAREST textures.
If the initial volume render texture is too big for the
WebGL renderer to handle, reduce it to half its size until
it can be managed by the renderer.
Also fix graphical glitches when zooming in too much on
surface plots.
Add title tags to the HTML files.