Changes to mobile viewing experience
For the 3D simulation the canvas should be at the top when in portrait mode.
Remove the planar slices view for iOS devices,
as it does not work here.
When on an Android device the default grid size is set to 64x64.Add a hovering div to the 2D version,
which notifies that the grid size has been reduced when on Android.

Also make the time for which to decrease the hovering div opacity for both
the 2D and 3D versions dependant on the time between frames.