Change default settings for the 2D simulation while on Android
Since only square domains are used in this simulation,
only the FFT shader intended for square domains was originally
used. It however did not work properly on Android, and it was
assumed that a CPU fallback was needed for this platform,
which was implemented and used in a previous commit.
It later turned out that the more general non-square shader does
work for this particular device configuration.
Since this works, the CPU FFT fallback is no longer used;
the GPU implementation is introduced back, and so the simulation
grid size no longer needs to be reduced.

Also added reverse bit sort CPU fallback to the fft3d script.
Here only the reverse bit sorting falls back to the CPU;
the full FFT CPU fallback was originally introduced in a previous
commit. Note that for the 3D simulation both the cubical and
non-cubical domain FFT shaders do not work while on Android, so
the full CPU fllback is needed here.